TheProject/requirements.txt
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
decorator==5.1.1
dnspython==2.2.0
email-validator==1.1.3
executing==0.8.2
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.29.1
greenlet==1.1.2
idna==3.3
ipython==8.0.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
mysql-connector-python-rf==2.2.2
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
prompt-toolkit==3.0.28
protobuf==3.19.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.31
stack-data==0.2.0
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.1.1
wcwidth==0.2.5
Werkzeug==2.0.3
WTForms==3.0.1
